int -> long

+++ src/main/java/com/example/bwengproj/repository/BidRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BidRepository extends CrudRepository <Bid, Integer> {
+public interface BidRepository extends CrudRepository<Bid, Long> {
 }
